Fix type mismatch
size_t isn't always an unsigned integer. Or an integer.

+++ sim_BuildROMs.c
@@ -89,7 +89,7 @@ struct ROM_File_Descriptor {
 #endif
 
 int sim_read_ROM_include(const char *include_filename, 
-                         int *psize,
+                         size_t *psize,
                          unsigned char **pROMData,
                          unsigned int *pchecksum,
                          char **prom_array_name,
@@ -221,7 +221,7 @@ return 1;
 }
 
 int sim_make_ROM_include(const char *rom_filename,
-                         int expected_size,
+                         size_t expected_size,
                          unsigned int expected_checksum,
                          const char *include_filename, 
                          const char *rom_array_name,
@@ -231,7 +231,7 @@ FILE *rFile;
 FILE *iFile;
 time_t now;
 int bytes_written = 0;
-int include_bytes;
+size_t include_bytes;
 int c;
 struct stat statb;
 const char *load_filename;
@@ -269,7 +269,7 @@ if (stat (rom_filename, &statb)) {
     return -1;
     }
 if (statb.st_size != expected_size) {
-    printf ("Error: ROM file '%s' has an unexpected size: %d vs %d\n", rom_filename, (int)statb.st_size, expected_size);
+    printf ("Error: ROM file '%s' has an unexpected size: %ld vs %zd\n", rom_filename, statb.st_size, expected_size);
     printf ("This can happen if the file was transferred or unpacked incorrectly\n");
     printf ("and in the process tried to convert line endings rather than passing\n");
     printf ("the file's contents unmodified\n");
